+++ src/js/api/hooks.js
@@ -27,11 +27,6 @@ export default function (UIkit) {
         this._initObserver();
 
         this._callHook('connected');
-
-        if (!this._isReady) {
-            ready(() => this._callReady());
-        }
-
         this._callUpdate();
     };
 
@@ -53,18 +48,6 @@ export default function (UIkit) {
 
         this._connected = false;
 
-    };
-
-    UIkit.prototype._callReady = function () {
-
-        if (this._isReady) {
-            return;
-        }
-
-        this._isReady = true;
-        this._callHook('ready');
-        this._resetComputeds();
-        this._callUpdate();
     };
 
     UIkit.prototype._callUpdate = function (e) {